refactor(schemas): move state and TypedDict schemas into dedicated package
- Added new package sp_oncall.schemas with:
  - state.py (GraphState, StepExecutionResult, ExecutedToolCall)
  - assessment_schema.py (AssessmentOutput)
  - planner_schema.py (PlannerOutput)
  - __init__.py re-exporting common types for simple imports
- Updated imports across codebase to use sp_oncall.schemas
- Left thin compatibility shims in original module locations to avoid breakage
