+++ src/sp_oncall/assessor.py
@@ -3,11 +3,11 @@ from typing import Dict, Any, Optional
 
 from langchain_core.messages import SystemMessage
 
-from sp_oncall.state import GraphState
+from sp_oncall.schemas import GraphState
 from sp_oncall.configuration import Configuration
 from sp_oncall.util.llm import load_chat_model
 from sp_oncall.prompts import OBJECTIVE_ASSESSOR_PROMPT
-from sp_oncall.assessment_schema import AssessmentOutput
+from sp_oncall.schemas import AssessmentOutput
 
 
 def objective_assessor_node(
