+++ src/sp_oncall/executor.py
@@ -1,10 +1,7 @@
 import asyncio
 from langchain_core.messages import HumanMessage
 
-from sp_oncall.state import (
-    GraphState,
-    StepExecutionResult,
-)
+from sp_oncall.schemas import GraphState, StepExecutionResult
 from sp_oncall.mcp_node import mcp_node
 from sp_oncall.prompts import NETWORK_EXECUTOR_PROMPT
 from sp_oncall.util.llm import load_chat_model
