+++ src/sp_oncall/schemas/planner_schema.py
@@ -1,0 +1,8 @@
+from typing import TypedDict
+
+
+class PlannerOutput(TypedDict):
+    """ """
+
+    objective: str
+    steps: str
